Sound recorder in C, using SDL2 and custom .wav header implementation.
It seems libsox lacks can't easily record sound form the default microphone:
https://stackoverflow.com/questions/44226188/libsox-record-from-default-microphone

The SDL2 library captures audio easily, though audio file formats are outside of it's scope.
It also has bindings to Haskell, plus I have some experience with it.

Currently recorded audio is distinguisable, but has crackling.
This is probably due to mistakes in endianness or encoding of some fields of the header.
For now, it should be enough to progress to python and then Haskell.